# perf(sourcemap): lazy compute decoded mappings