--- a/src/rollup/types.d.ts
+++ b/src/rollup/types.d.ts
@@ -52,7 +52,7 @@ export type SourceMapSegment =
 
 export interface ExistingDecodedSourceMap {
 	file?: string;
-	mappings: SourceMapSegment[][];
+	readonly mappings: SourceMapSegment[][];
 	names: string[];
 	sourceRoot?: string;
 	sources: string[];