--- a/src/utils/collapseSourcemaps.ts
+++ b/src/utils/collapseSourcemaps.ts
@@ -6,6 +6,7 @@ import type {
 	LogHandler,
 	SourceMapSegment
 } from '../rollup/types';
+import { decodedSourcemap, resetSourcemapCache } from './decodedSourcemap';
 import { LOGLEVEL_WARN } from './logging';
 import { error, logConflictingSourcemapSources, logSourcemapBroken } from './logs';
 import { basename, dirname, relative, resolve } from './path';
@@ -150,7 +151,7 @@ class Link {
 
 function getLinkMap(log: LogHandler) {
 	return function linkMap(source: Source | Link, map: DecodedSourceMapOrMissing): Link {
-		if (map.mappings) {
+		if (!map.missing) {
 			return new Link(map, [source]);
 		}
 
@@ -224,6 +225,10 @@ export function collapseSourcemaps(
 
 	sourcesContent = (excludeContent ? null : sourcesContent) as string[];
 
+	for (const module of modules) {
+		resetSourcemapCache(module.originalSourcemap, module.sourcemapChain);
+	}
+
 	return new SourceMap({ file, mappings, names, sources, sourcesContent });
 }
 
@@ -246,5 +251,5 @@ export function collapseSourcemap(
 		getLinkMap(log)
 	) as Link;
 	const map = source.traceMappings();
-	return { version: 3, ...map };
+	return decodedSourcemap({ version: 3, ...map });
 }