<!--
  ⚡️ katchow! We ❤️ Pull Requests!

  If you remove or skip this template, you'll make the 🐼 sad and the mighty god
  of Github will appear and pile-drive the close button from a great height
  while making animal noises.

  Pull Request Requirements:
  * Please include tests to illustrate the problem this PR resolves.
  * Please lint your changes by running `npm run lint` before creating a PR.
  * Please update the documentation in `/docs` where necessary

  Please place an x (no spaces - [x]) in all [ ] that apply.
-->

This PR contains:

- [ ] bugfix
- [ ] feature
- [ ] refactor
- [ ] documentation
- [x] other

Are tests included?

- [ ] yes (_bugfixes and features will not be merged without tests_)
- [x] no

Breaking Changes?

- [ ] yes (_breaking changes will not be merged unless absolutely necessary_)
- [x] no

List any relevant issue numbers:

<!--
If this PR resolves any issues, list them as

  resolves #1234

where 1234 is the issue number. This will help us with house-keeping as Github will automatically add a note to those issues stating that a potential fix exists. Once the PR is merged, Github will automatically close those issues.

If an issue is only solved partially or is relevant in some other way, just list the number without "resolves".
-->

### Description

By deferring the decoding of the sourcemap mappings field, we can avoid extra computation and garbage generation, both of which appear to save a significant amount of time.

In my test application (using vite), I saw a reduction in total vite build time of about **28.4%** _(8.76s)_, and a reduction in peak heap size of **19.6%** _(706 MB)_. [Update] The latest commit reduces the peak heap size by 36.6% in total (1,300 MB) for projects that deal with large sourcemaps. While the first commit reduces allocations, the second commit technically allocates more, but frees up more data to be garbage collected which has a net positive effect. The stats below reflect only the first commit, a summary of the two commits combined can be found in this comment: https://github.com/rollup/rollup/pull/5075#issuecomment-1656918949

Project Info:
* Node: v18.17.0
* Vite v4.4.7
* Rollup v3.26.3
* Module Count: 5490 modules transformed
* Chunk Count: 54 chunks rendered

Before:
* Average vite build time (5 runs): 30.89 seconds
* Average peak heap size (5 runs): 3,594.83 MB
* Time spent in GC: ~13.7 seconds
* Time spent in `decode()`: ~3.11 seconds

After:
* Average vite build time (5 runs): 22.13 seconds
* Average peak heap size (5 runs): 2,888.2 MB
* Time spent in GC: ~5.9 seconds
* Time spent in `decode()`: ~1.70 seconds

I'm happy to try to write tests for this if it is needed, but as I'm still getting my feet wet in this project that might take me a little while to figure out exactly how to go about it.

I'm also a little curious if there is a higher-level optimization to be applied that would make this no longer necessary. For example, what are the cases where we are calling `decodedSourcemap`, but then abandoning the result without ever actually accessing the mappings value? 

Are we able to avoid call to `decodedSourcemap` at all and/or any json parsing until we know we will need it to be decoded? 

Are there cases where we decode the sourcemap only to pass it back into the SourceMap constructor which then re-encodes it to a string? 

While I think those are all good questions to ask, the current solution provides enough benefit and seems relatively low risk, so IMO its worth pursuing on its own.

Allocations Before:
<img width="315" alt="image" src="https://github.com/rollup/rollup/assets/17907922/df3b4960-b7e3-4d4a-89b0-f7a55e1a3597">

Allocations After:
<img width="419" alt="image" src="https://github.com/rollup/rollup/assets/17907922/8fca4313-dcbd-4a6f-bf55-522291a22c19">