@thebanjomatic 2023-07-27
@lukastaegert Let me know if you can take a look at this. I started down this path because I was running OOM during a build on CI and I'm hoping this change can make a dent in improving https://github.com/vitejs/vite/issues/2433