@thebanjomatic 2023-07-28
Review has been updated based on feedback.

Adding on another datapoint, I tested: https://github.com/Axeldeblen/realworld-big-build-vite-4

The default configuration doesn't build sourcemaps, but we can still see some improvements
Before: ✓ built in 2m 9s
```js
{
  stats: {
    peakMemory: 4_148_762_536,
    decodedSourcemapCalls: 16133,
    decodeCalls: 16133
  }
}
```

After: ✓ built in 1m 59s
```
{
  stats: {
    peakMemory: 3_926_596_880,
    decodedSourcemapCalls: 16133,
    decodeCalls: 0,
    memoizedCalls: 0
  }
}
```

Unfortunately, if I enable sourcemaps in the config, its a bit more of a mixed bag. For this configuration we only really save 1 call to `decode()`, and otherwise just introduce some overhead to each access of `.mapping`. That being said, this project isn't really consuming any dependencies with sourcemaps and is mostly javascript+svelte, so its not the best example of where this sort of change would shine, but it is disappointing to see that it's a little slower for that case.

Before: ✓ built in 2m 17s
```js
{
  stats: {
    peakMemory: 4149991032,
    decodedSourcemapCalls: 16139,
    decodeCalls: 16139
  }
}
```

After: ✓ built in 2m 24s
```js
{
  stats: {
    peakMemory: 4150411624,
    decodedSourcemapCalls: 16139,
    decodeCalls: 16138,
    memoizedCalls: 14130,
  }
}
```