@thebanjomatic 2023-07-29
@lukastaegert I have iterated on this idea a little further and have some additional gains. For a complete description, see the second commit: https://github.com/rollup/rollup/pull/5075/commits/c0c0743d9cf131f31aaa0469fcbf3e7bc74be41c

This change improves on the peak memory usage. I did some binary searching in both my internal enterprise app, and also in https://github.com/Axeldeblen/realworld-big-build-vite-4 to identify the lowest `--max-old-space-size` value (to the nearest 50MB) that still builds successfully.

Internal Vue app:
* master: 3550 MB (29.22s average)
* prior commit: 2650 MB (21.17s average)
* latest commit: 2250 MB (20.87s average)

realworld-big-build-vite-4 (+ sourcemaps: true):
* master: 4100 MB (~137s average)
* prior commit: 4100 MB (~136s average)
* latest commit: 3900 MB (~128s average)

realworld-big-build-vite-4 (default config w/ sourcemaps: false):
* master: 4000 MB (~133s average)
* prior commit: 3850 MB (~118s average)
* latest commit: 3800 MB (~118s average)