@thebanjomatic 2023-07-29
The only minor issue I have with this PR that could possibly be solved by a breaking change is that the `RollupCache` type contains `ExistingDecodedSourceMap` and `DecodedSourceMapOrMissing[]`. I know rollup doesn't currently persist the cache to disk, but this could be problematic because serializing the sourcemaps with `JSON.stringify()` will re-hydrate the decoded array in the process. This is a little undesirable both from a cache file-size perspective and it seems like it would be more natural to store these as `ExistingRawSourceMap` in the cache, and then construct a `decodedSourcemap` from the cached module for things like watch builds, etc. Having sourcemaps stored in encoded form in the cache would force you to call decodedSourcemap to get the appropriate types, so it makes it easier to write the correct code.