@thebanjomatic 2023-08-02
Thanks for the review @lukastaegert I have pushed a commit dropping the setter. This is good to merge from my end when you are ready.